CRAS: alsa_mixer - Add coupled mixer controls for speaker
Let cras_alsa_card queries the coupled mixer names from UCM and pass an array
of coupled mixer names to cras_alsa_mixer_create.
Add coupled_mixer_control struct in cras_alsa_mixer to handle a group of
controls that should be changed together for their volume and mute/unmute state.

Note that there are some assumptions made here because of lack of
fully-specified UCM support as we want to make the minimum change to be
merged to M49.
- A virtual mixer_control will be created as the base of mixer_output_control.
- The mixer_output_control is associated with speaker node only.

In the future changes, these will be added:
- The coupled controls will be used in volume change and mute/unmute.
- cras_alsa_io should handle the speaker node creation for virtual
  mixer_control.

When fully-specified UCM is added, the code path in
cras_alsa_card_create and cras_alsa_mixer_create should be removed.

BUG=chrome-os-partner:48176
TEST=make check

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Iad86681f8e883c650433677063af81ec3111d992
Reviewed-on: https://chromium-review.googlesource.com/323952