CRAS: alert - allow data arguments to alerts
Add the ability for an alert to be signaled with an optional data
parameter.  Alerts triggered in this way can behave differently.
Optionally if they are signaled N times with data, they will be
called N times.  By default similar to normal alerts without
data, those with data will combine alerts if they happen rapidly
together, possibly only executing the callback once with the most
recent data pointer.

This will be used to allow data to be passed in to the callbacks
so that more callback systems can use the alert infrastructure
instead of open-coding it.

BUG=None
TEST=Unit tests added. Used via cras client monitor in subsequent
     commit.

Change-Id: I96c5a0485dca8b16dbb0f45d3e5114ca1b491abb
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/348653
Commit-Ready: John Muir <muirj@google.com>
Tested-by: John Muir <muirj@google.com>