CRAS: iodev - Fix a bug internal speaker got disabled
Commit e53b58ef causes a bug that internal speaker got
disabled when headphone unplugs, due to both nodes are
associated to the same iodev.
This bug is hard to notice because Chrome always set internal
speaker back to active immediately, but a few samples are lost.
Fix this issue by checking if the unplugged node is the active
node of the iodev it belongs to.

BUG=chromium:586026
TEST=On Peppy, 'stop ui' to better observe this issue. Plug
then unplug 3.5mm headphone jack, 'cras_test_client --dump_s'
to see internal speaker is still active.

Change-Id: Ia48d010cf095b67266871d7c7b929b60e8c8f89d
Reviewed-on: https://chromium-review.googlesource.com/330293
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>