UCM config: glados and chell: Set Mic Volume to 24dB
The default value of "Mic Volume" control is 0dB which
is too low. This patch sets the gain to +24dB.

This patch also binds the "Mic Volume" control to the
sklnau8825adi Headset Jack, so that CRAS can adjust it
at runtime for AGC.

BUG=chrome-os-partner:51151
TEST=The remote side can hear clearly in hangouts.
See "cras_test_client --capture_gain ..." affects "Mic Volume".

Change-Id: If0cf259e30e20a79b7c61776faed79e39c77ab3c
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/332645