CRAS: alsa_ucm : Add flag to enable software gain
There are boards without any gain control for internal microphone.
To apply software gain on this node, user can specify MaxSoftwareGain "2000"
in SectionDevice for internal microphon.
The system capture gain limit will be -5000 to 2000 dBm.
Note that if there is capture gain control in mixer or on the node,
then the combined hardware gain will be set to 0dB if software gain is
used.
The following patch series will implement the above feature.

BUG=chrome-os-partner:50139
TEST=make check

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I814efec846140dfec75e51967017fe5473e8b9c3
Reviewed-on: https://chromium-review.googlesource.com/332914
Reviewed-by: Ben Zhang <benzh@chromium.org>