CRAS: iodev - Add start function to cras_iodev
In alsa_io, simplify dev_running to just check if device is running.
In alsa_io, implement start function to start PCM.
In audio thread, start device when sample is ready from stream and is
written to the deivce.

Avoid putting zeros to device when it is still waiting for the first reply
from stream in these two cases:
- If there is no stream on the device, only fills zeros to device when
device is running. Only device in idle stage will get these zeros.
- If there is no sample written, and device buffer is empty, only fills
zeros to device when device is running.

Add dummy start function definition to differnet iodevs.
For bluetooth and loopback devices, the devices start running when they
are opened.
For empty device and test device, devices are always running.

BUG=chromium:519942
TEST=cras_test_client --playback_f
/usr/local/autotest/cros/audio/fix_440_16.raw --show_latency; check the
latency is reduced from 13ms to 4ms on headphone, from 19ms to 10ms on
speaker.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I0b54e8e78d6dd64e0e92d6c140fe644385483c9b
Reviewed-on: https://chromium-review.googlesource.com/333837