CRAS: Allow ALSA period_frames to be specified.
Some devices need to setup the period_frames before audio is written
to the mmap in order to configure down-stream devices such as a DSP
that pulls the data from the memory map based on a timer.

This change allows the period_frames to be specified per io-device,
by adding a new UCM configuration parameter. Configuration of the
period frames is done within the normal ALSA hwparam setup.

BUG=None
TEST=When MinPeriodFrames is not used, nothing changes.
     When MinPeriodFrames is defined, the value is modified within
     the device's ALSA hwparams.
     Add unit tests to cover the new configuration.

Change-Id: Ia2d6352d063a0d2692d05ba06be40b010bf59b26
Reviewed-on: https://chromium-review.googlesource.com/347361
Commit-Ready: John Muir <muirj@google.com>
Tested-by: John Muir <muirj@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>