CRAS: bt_device - Fix bug in profile switching
cras_bt_device exposes functions for profile switching at iodev
open and at close. The reason is that when bt iodev is about to
open and meets the condition to switch profile, main thread needs
to re-enable this iodev after profile switching is done.
In addion to open/close, when each profile(A2DP or HFP) is connected
profile switch could happen but we should leave the enable/disable
status of bt iodev unchanged, or it could accidentally be enabled
while other device is in use.

This change fixes the bug that BT headset got enabled while internal
speaker is active. Also the profile switching functions are renamed
from
  - cras_bt_device_switch_profile_on_open()
  - cras_bt_device_switch_profile_on_close()
to
  - cras_bt_device_switch_profile_enable_dev()
  - cras_bt_device_switch_profile()

so it is more clear in which scenario we need to do the extra task
to enable the dev.

Note that this bug wasn't discovered until latest release because Chrome
always selected BT headset for active output when it is connected. And
now the behavior is changed to auto-select BT headset based on user's
last preference.

BUG=chrome-os-partner:54176
TEST=Connect BT headset, play audio and select output to internal
speaker. Restart system, seeing that internal speaker is the active
output. Play audio and now try connect BT headset, verify that when
BT headset is connected it doesn't play audio.

Change-Id: If30088a7bbe5d2b8caa726bce7feb79f73f44d31
Reviewed-on: https://chromium-review.googlesource.com/354793
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>