CRAS: Libcras auto-reconnect to server.
Implement automatic reconnect to the audio server that works if
the audio server is stopped and restarted an arbitrary amount of
time driven by the control/command thread.

Re-implement the cras_client_connect functions such that they
will wait the timeout period specified even if the server is not
yet running.

When re-connecting to CRAS, all observation/notification callback
functions are re-registered with the server.

Add a connection status callback function which is called to
indicate the status of the connection to the server.

Clarify the API initialization sequence, and provide a method to
initialize and initiate a connection to CRAS with non-blocking
functions. Add a helper that does the new recommended sequence.

Clarify the API header to define which functions require that the
control/ command thread is running, and which functions may block
waiting for a connection to the audio server.

BUG=None
TEST=All cras unitttests pass.
     - cras_monitor can continue to execute survive server
       connection cycle.
     - cras_monitor can start using fully async functions.
     - cras_monitor's connection callback is executed when the
       server connection is established or drops.
     - Ensure that libcras and CRAS continue to function
       correctly on samus.

Change-Id: I60c31b86bc00df33aa2c2ebcf0df6e76aa01eb7c
Reviewed-on: https://chromium-review.googlesource.com/359938
Commit-Ready: John Muir <muirj@google.com>
Tested-by: John Muir <muirj@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>