CRAS: audio_thread - Align stream fetch time
A usage pattern is found to make output buffer drift crazily high.
While output device running with stream A, the buffer level is
expected to stay in range [X, X + d1]. Suppose the new stream B is
connected when output buffer at level X + d' and stream A is
immediately removed, the output buffer will change to stay in
[X + d', x + d' + d2].
The root cause is that when a new stream is added when output device
is already running, its fetch time isn't aligned with any of the
existing stream, so the above scenario is very likely to happen.

BUG=chromium:623868
TEST=Use audio_CrasOutputStress test to verify the problem is fixed.

Change-Id: I071c76ea647a587fb7bd6e8361eedf1029474fc8
Reviewed-on: https://chromium-review.googlesource.com/360880
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>