CRAS: libcras: Use rwlock to access server_state data.
The server_state pointer is a memory map to a shared memory
region created by CRAS. If CRAS dies, then this pointer will be
set to NULL, and the memory map will be closed. Ensure that there
is no race condition between the readers and writers of this
pointer and the memory it points to.

Use an external structure for the lock.

BUG=None
TEST=Exercise all functions that are modified.
     Unit tests pass.
     Run on samus.

Change-Id: I5d6ae3f75ed19be2fdd7aac2518c3eca7efa2942
Reviewed-on: https://chromium-review.googlesource.com/360950
Commit-Ready: John Muir <muirj@google.com>
Tested-by: John Muir <muirj@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>