CRAS: alsa_io - Get volume curve for node by UCM
This changes allows spcifying volume curve for a node associated
to specific UCM device, which is useful in the case that 'Line Out'
and 'Headphone' nodes shares the same gpio jack and mixer control.

BUG=chromium:634958
TEST=Edit CRAS config under /etc/cras and specify the section label
as the UCM device of a node to verify volume curve applied.

Change-Id: Icec8de79385423922b9a0d52942cee833c8a07cc
Reviewed-on: https://chromium-review.googlesource.com/367682
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>