scripts: Fix audio-tuning frontend with new WebAudio biquad filter implementation
The spec of WebAudio biquad filter "lowpass" and "highpass" has
been changed. Chrome's implementation was changed in 53.0.2768.0.
The new implementation uses the Audio EQ Cookbook formula, different
from the original one, but still expects Q in dB.
See https://github.com/WebAudio/web-audio-api/issues/771

This patch adds detection of the filter implementation. If the
browser uses the new formula, conversion of Q is made to
simulate the original filter frequency response with the new
formula. This ensures the audio-tuning playback is always
consistent with the current CRAS DSP implementation, on any
browser verion.

BUG=chrome-os-partner:55365
TEST=DRC/EQ testing on Chrome before/after the filter change

Change-Id: I3e67a9b87f94b228d173fa08dcaa0fe4df5ac64e
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/367551
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>