CRAS: hfp_ag - Delete a2dp delay timer
This is a preparation for future change to handle the case where
HFP/HSP connected but A2DP failed. Delayed callback waiting for
A2DP connection was used as a workaround to avoid user hear BT
audio transition between profiles. But it turns out not good enough
so delete it and replace by a better mechanism later.
This change adds additional layer when HFP SLC is initialized and
notify bt_device to actually start the HFP audio gateway. This is
a preparion for later commits to defer starting HFP audio until
A2DP is ready.

BUG=chromium:543444
TEST=None

Change-Id: I269d3b459f0e575ab1e0e3021cb531580a285562
Reviewed-on: https://chromium-review.googlesource.com/377844
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>