CRAS: bt_device - Add connection watch timer
For a BT headset supporting more than one profile, we should not
assume the headset handles profile connection perfectly and just
wait for each profile to be ready.
This change adds the responsibility to CRAS to catch cases that
any profile fails or forgets to connect. A timer is used since
BT device change to 'connected' state and watches the readiness
of each supported profile.

BUG=chromium:543444
TEST=Use Motorola S10-HD headset, press the 'call' button while
it's not connected. Verify that both HFP and A2DP will connect.

Change-Id: Icec49d6b1c6d437fac8aaa0ffb6a90f7a3cd3020
Reviewed-on: https://chromium-review.googlesource.com/377845
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>