CRAS: bt_device - Start A2DP and HFP when both are ready
There are cases that a headset connects A2DP but not HFP/HSP.
Modify the connection watch logic in bt_device to check and wait
for every supported profile(HFP, A2DP). If certain profile is
not ready in a while, try connect it.

BUG=chromium:543444
TEST=Use Motorola S10-HD headset, press the play/pause button
while it's not connected. Verify that both HFP and A2DP will
connect.

Change-Id: I4d729dc4e7c6da1a78cf8e748b7b17ee9912f3d3
Reviewed-on: https://chromium-review.googlesource.com/377847
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>