CRAS client: Start stream audio thread before connection
When the server sends the first data request to the client, it is
important that the client is ready to respond instantly. This
change ensures that this is the case by starting the stream's audio thread
before the connection to the server has been established. If
there is a problem with the server then the audio thread is
stopped as per normal.

BUG=None
TEST=Unit tests updated.
     Audio playback works on new device.
     Audio playback and capture work on Samus.

Change-Id: I0b1d96457d98601643da6516c0cf15ff2af351b1
Reviewed-on: https://chromium-review.googlesource.com/387135
Commit-Ready: John Muir <muirj@google.com>
Tested-by: John Muir <muirj@google.com>
Reviewed-by: Dylan Reid <dgreid@chromium.org>