CRAS: alsa_helper - Fix num_underrun count
When snd_pcm_avail returns a number that equals buffer size, it should
be counted as underrun as well although it might not be an underrun if
total_written happens to be greater than zero.

This is because audio_thread calls cras_iodev_output_underrun when
hw_level and total_written are both zero. If we do not count the case
where avail == buffer_size (that is, hw_level == 0) as an underrun,
num_underrun dumped in audio thread log will be 0, while there is delay
caused by cras_iodev_output_underrun.

With this change, we can rely on num_underrun == 0 in audio thread log
to examine that hw_level remains high enough.

BUG=chromium:667674
TEST=check num_underruns is the same as number of times
cras_iodev_output_underrun is called.

Change-Id: I170753485b135e5170ff810830668a1b03ece04f
Reviewed-on: https://chromium-review.googlesource.com/425054
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>