CRAS: iodev - Do not ramping up when output should mute
When output should be muted, do not ramping up for sample being ready.
This is for the case where stream is added and removed while system
remains muted.

BUG=chromium:669662
TEST=make check
TEST=Let system remain muted, play sine tone and stop repeatedly.
Unexpected audio happens only in the first time, and system remains
silence for all other trials. The unexpected audio will be fixed in the
next CL.

Change-Id: I55cd544faac42cf1ac72b9b5445bc7a5c18a3a05
Reviewed-on: https://chromium-review.googlesource.com/428730
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Chinyue Chen <chinyue@chromium.org>