CRAS: iodev_list - Skip ramping if device volume is zero
When adjusted active node volume is zero, or system volume is zero,
samples of a device will be suppressed to zero.
Since the device is already playing zeros, there is no need to ramp down
or up for mute/unmute change.

BUG=chromium:697511
TEST=on kevin, check there is no extra ramp when pressing volume down
when volume is 0.
TEST=make check

Change-Id: I648fab83f2e6dcfbd2db5370db0509d513d3b40d
Reviewed-on: https://chromium-review.googlesource.com/447852
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>