CRAS: iodev_list - Fix closing pinned stream
The expected suspend/resume flow:
When system suspends, all the streams are disconnected from device, and
are kept in stream_list. After resume, the device will be opened, and
streams will be attached to devices.

When closing a device, we should check whether there is a pinned stream
currently attached to the device, instead of checking whether there is a
stream that should be pinned to device.
Otherwise, at suspend we will not close the device which has played
a pinned stream. This breaks the flow for resume where we skip opening
the device since it is in open state, while in fact it is in suspended
state.

BUG=chromium:748201
TEST=make check
TEST=On auron_paine, play audio using pinned stream:
 $ cras_test_client --playback_f
/usr/local/autotest/cros/audio/fix_440_16.raw --pin 8
 $ powerd_dbus_suspend
 Then wake up system and make sure audio keeps playing.
Change-Id: I4cfe020db79bee21135df27da81042b98a83dbbc
Reviewed-on: https://chromium-review.googlesource.com/585033
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>