Plankton Captured HDMI Audio ucm-config for gain=0
The Plankton HDMI capture chip has a gain range of -16--+12dB,
which leads Cras to set the capture gain to +12dB by default.
Unfortunately a gain > 0 leads to a distorted sound for input
with the volume maximized.

By specifying FullySpecifiedUCM=1 in the UCM file without
specifying a mixer, the capture gain stays at 0--0dB and the
sound is not distorted.

BUG=b:66932284
TEST=Verifyed by checking the output of cras_test_client +
	ensuring the sound is not distorted through listening.

Change-Id: I40c892d65bef9afb648a43e79809c7bf5ff60ca8
Reviewed-on: https://chromium-review.googlesource.com/743726
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Tove Petersson <tovep@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>