CRAS: rstream - Let rstream user query pending reply
Export interfaces to let rstream user query whether this rstream is
pending a reply from client.

Previously, this pending flag is maintained in cras_shm, and the user of
rstream sets/clears/queries it directly. This will be changed in the
following patch series such that

- cras_rstream maintains the state internally using flag in cras_shm.
- dev_stream talks to cras_rstream, and provides a thin layer for dev_io.
- dev_io talks to dev_stream.

A stream is pending a reply from client after it sends an audio message
to client (request audio for playback, or or audio ready for capture).

Let rstream handle this state change so dev_stream does not need to set
the flag itself.

BUG=b:68479798
TEST=unittest to be added

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: I6600ae94e38975cc49b92769528a97ddb73e72ce
Reviewed-on: https://chromium-review.googlesource.com/766107
Reviewed-by: Dylan Reid <dgreid@chromium.org>