CRAS: dev_stream - Capture stream should know whether it is pending
Hotword stream needs to decide whether it should schedule a wake.
It does not need to schedule a wake if it is pending the reply from
client, because the reply from client will wake audio thread.

In get_input_wake_time, there was a bug that audio thread falsely believes
that client has received the request from server if hw_level is larger than
cb_threshold.

However, this is not true. Consider a scenario:

1.In capture_to_streams: hw_level is still less than cb_threshold. Audio
thread does not send data ready request to client.

2.In get_input_wake_time, hw_level is more than cb_threshold. Audio
thread thought it has already sent the request to client, while it did
not.

To fix this, dev_stream needs to query whether the stream is pending a
reply from client.

Also, fix dev_stream_poll_stream_fd, such that the fd is added to ppoll
if stream is pending the reply from client.

Detail in b/68479798.

BUG=b:68479798
TEST=unittest
TEST=1. On eve, cras_test_client --listen /tmp/rec.raw.
     2. powerd_dbus_suspend
     3. "Ok google" to wake up device.
     4. cat /proc/asound/card0/pcm5c/sub0/status and see device is in
     RUNNING state, not PREPARED state.
     5. There will be no data recorded. The following patch will fix
     this.
Change-Id: I77b754ec1c33483ddf7e857d39b8dde825f4a0e5
Reviewed-on: https://chromium-review.googlesource.com/763948
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>