CRAS: audio_thread - keep calling read until all bytes are read.
This is a speculative fix for a fuzzer crash in
cras_fmt_conv_destroy. It looks like the returned struct
cras_fmt_conv is invalid. It can happen if read returns a
number less than requested. Add a loop to read until finished
to see if the crash will be fixed.

According to man page:
Read Return Value
"It is not an error if this number is smaller than the number
of bytes requested; this may happen for example because fewer
bytes are actually available right now (maybe because we were
close to end-of-file, or because we are reading from a pipe,
or from a terminal), or because read() was interrupted by a
signal.

BUG=b:68495404
TEST=Play left/right speaker test on YouTube on cave.
"cras_test_client --config_global_remix 2:0.5,0.5,0.5,0.5"

Change-Id: I9441c9bb0f6c9750b3d56960bbab8d480d3f0a32
Reviewed-on: https://chromium-review.googlesource.com/784711
Commit-Ready: Wu-Cheng Li <wuchengli@chromium.org>
Tested-by: Wu-Cheng Li <wuchengli@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>