CRAS: Run restorecon on shm fds
CRAS stream FDs need to have the correct SELinux label in order for the
ARC audio HAL cras client to access them. This is required beginning
with Android O, because the audio HAL runs under Treble as a separate
process, and thus it has a more restrictive security policy than in
earlier versions of Android.

This patch accomplishes that, buy deriving the real filepath of an
fd from procfs, and then running restorecon on it.

BUG=b:68722835
CQ-DEPEND=CL:851834
TEST=Stream fds can be opened successfully on the audio HAL.

Change-Id: I4792c968eca509a93cf9edf6b58aeceef663f3b4
Reviewed-on: https://chromium-review.googlesource.com/851740
Commit-Ready: Prashant Malani <pmalani@google.com>
Tested-by: Prashant Malani <pmalani@google.com>
Reviewed-by: Luis Hector Chavez <lhchavez@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>