CRAS: Add system-level audio active state.
Adds a binary flag to system state that denotes whether CRAS currently
has any streams that are active, and also playing/capturing audio that
is not empty (just zeros).

This will be used by powerd to allow the device to sleep if there are
active audio streams, but those streams are not actually producing any
sound (for example, a misbehaving ad creative).

BUG=chromium:753596
TEST=Ran locally on eve

Change-Id: Ie56a4460a6e56bbb757d866410717b98770b97ea
Reviewed-on: https://chromium-review.googlesource.com/888185
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>