CRAS: Add dbus interface for active status
The signals added here will be used to tell powerd when the state
changes, so that it knows whether it can put the device to sleep.

The method added to the interface will be used by powerd at startup to
query the current status.

BUG=chromium:753596
TEST=Ran locally on eve

Change-Id: Icff42f7d71ea7bede2187a8d74682f1fad4b6178
Reviewed-on: https://chromium-review.googlesource.com/888186
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>