CRAS: audio_format - Support mapping RL and RR channels to SL and SR.
Currently, a stream with SL and SR channels can play on a device with
RL and RR channels (but no SL/SR channels) - and the streams will be
mapped accordingly. The inverse (a stream with RL/RR playing to a
device with SL/SR) is not possible. This change enables that inverse
mapping.

BUG=chromium:831088
TEST=Ran locally on eve

Change-Id: I7613249e937af760449d0f4699de47d2f6cbe804
Reviewed-on: https://chromium-review.googlesource.com/1004814
Commit-Ready: Louis Collard <louiscollard@chromium.org>
Tested-by: Louis Collard <louiscollard@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>