CRAS: dbus - api to query if aec is supported
This api will be used by Chromium to know if system aec
can be used instead of the one in browser.
Set system aec non-supported by default and use board config
to enable on specific boards for initial testing.
In future this should be set to 'supported' by default and
only disable on specific boards when needed.

BUG=chromium:710465
TEST=Edit board.ini to set system aec supported flag and
then use dbus-send to verify the correct value is returned.

Change-Id: Ib03c9b9d32d87dab8a3eaf1a14d27a0d8f51a143
Reviewed-on: https://chromium-review.googlesource.com/1111896
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>