CRAS: fmt_conv - Bug fix in cras_fmt_conv
When re-calculating the room in output buffer at SRC and linear
resampling, the intermediate limit was determined using
cras_frames_at_rate() function, and changed in commit 4aca14a79
to take the floor instead of ceiling of calculated number.

There is a bug found when limit count is smaller than the ratio
of in/out sample rate (e.g 16K and 48K), the intermediate limit is
rounded to zero and that confuses both the linear resampler and
rate converter. Fix this issue by making sure the limit value is
at least 1.

BUG=chromium:860408
TEST=unittest

Change-Id: I64a401a684ca9c694c5fa40a7fa6e79883e15ba2
Reviewed-on: https://chromium-review.googlesource.com/1126800
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>