CRAS: iodev - Handle input_dsp_offset larger than read frames
input_dsp_offset is used to track to where in input buffer we've
applied dsp. Handle the case when number of read frames is smaller
than this offset value.

Note that this "input_dsp_offset >= frames" case didn't happen before
we add APM processing. The scenario to trigger this is by creating an
input stream of buffer size smaller than 10 ms equivalent of data,
which is the chunk size APM requires when processing data.

The exact detail is complicated when problem triggered. The rough idea
is that: suppose APM processes data in chunk of N frames which is larger
than stream shm size M.  At some point audio thread could read N frames of
data from device, which makes input_dsp_offset = N. But stream side has
only M frames of room, so the number of requested/read frames from device
is M.

It's very difficult to avoid the "input_dsp_offset >= frames" case.
One possible solution to avoid this is to rework the dev_stream_capture_avail()
to report something smarter.  But that doesn't work beceause of a
foundamental issue: when APM's buffer is full, the stream capture avail
has to report '0'.  And that will confuse audio thread with the case when
client is late to consume shm data.

BUG=chromium:863284
TEST=cras_test_client --capture_file /tmp/1 --effects aec --block_size 170
And use valgrind to verify that there's no memory issue.

Change-Id: I6bca034b5c1fbff9d00cdb637015249953133d43
Reviewed-on: https://chromium-review.googlesource.com/1141680
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>