CRAS: Add seccomp policy file
In order to integrate webrtc-apm library in CRAS, we want to
make CRAS run in an environment as close to how Chrome renderer
process is sandboxed.
This change adds seccomp.policy file which will be installed in
ebuild.

BUG=chromium:857378
TEST=Apply whole patch set including ebuild changes
Test on Eve and Kevin that seccomp policy files are installed
and used.

Change-Id: Ieb74924634bd6e970c250f78b54b2d19b6e90f87
Reviewed-on: https://chromium-review.googlesource.com/1136345
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>