CRAS: fix ramping for large block size stream
Ramping takes an initial scaler and an increment.
When the block size is large, the incremented scaler will be greater
than software volume scaler.
This causes a pop noise in the beginning of playback.

To fix this, we need to pass an upper limit to the scale ops.

BUG=b:113041524
TEST=unittest
TEST=run cras_test_client --playback_f
/usr/local/autotest/cros/audio/fix_440_16_half.raw --block 4800
there is no noise in the beginning.
Change-Id: I2f2186d6af64b015335e26d093984ca737cadbf9
Reviewed-on: https://chromium-review.googlesource.com/1270329
Commit-Ready: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>