CRAS: cras_iodev - Let audio thread wake up earlier to avoid underrun
When the client stream buffer is larger than hardware buffer, we need to
write it separately into device. In current version, the audio thread
will wake up when hw_level drops to 1ms. It is easy to trigger underrun.
So we change it to wake up when hw_level drops to min_cb_level if the
frames are ready to be played.

BUG=b:117915490
TEST=aplay works fine on grunt.

Change-Id: Ied4555e470d5ca973fb5a685f59a39aaeabeb645
Reviewed-on: https://chromium-review.googlesource.com/1322291
Commit-Ready: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Reviewed-by: Dylan Reid <dgreid@chromium.org>