CRAS: Do not mix stream before its first fetching
When we add the new stream, the next_cb_ts of this stream would be set
to the earliest next_cb_ts of others. It causes the new stream keep
empty before next fetch time. It also means we can not write any frames
during this time. However, if we use the big client buffer and the
small device buffer, the next fetch time is far from now. It will
trigger underrun.

Hence, we need to ignore new stream until it starts fetching and
running. In order to do it, we create new stream list "pending_streams"
for each device. First, we add the new streams in this list. And then move
streams from pending_streams to the original stream list when they request
fetching. So the new streams will not affect other streams anymore.

BUG=b:117915490
TEST=Playing two streams does not trigger underruns on grunt.

Change-Id: I4503dff5a276d2c63c64476d417a7cc5e03f483b
Reviewed-on: https://chromium-review.googlesource.com/1341708
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>